requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-dotenv==1.1.1
httpx==0.28.1
pydantic==2.12.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
chromadb==1.0.20
langchain==0.3.27
langchain-community==0.3.29
elevenlabs==2.18.0
twilio==9.8.4
